requirements.txt
# pip install -r requirements.txt

numpy==1.19.5
opencv-python==4.5.4.60
opencv-contrib-python==4.5.4.60
opencv-python-headless==4.5.4.60
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.0
PyYAML==5.3.1